FileWatching: ensure PollingFileWatcher reports StatStruct correctly
Previously, we might miss an prev=>current update event while waiting in the workqueue.
Now a careful application can ensure it always has the most up-to-date stat,
without needing to do an extra call to Stat after the notify.
Obviously, you still must be careful to initialize this before attempting the initial
operation (to ensure the previous Stat is the intended one).